:recycle: (UIExplorer): RemoteArrow - refactor buttons
- use SF symbols for arrows as asked by design
- update init
- remove unused TriangleShape
- rename button view
- fix touch zone of button
