--- Apps/LekaActivityUIExplorer/Sources/GameLayoutTemplates/Components/AnswerLayouts/Layouts/Remotes/RemoteArrow/ArrowButtons/Components/TriangleShape.swift
@@ -1,19 +1,0 @@
-// Leka - iOS Monorepo
-// Copyright 2023 APF France handicap
-// SPDX-License-Identifier: Apache-2.0
-
-import SwiftUI
-
-struct TriangleShape: Shape {
-    func path(in rect: CGRect) -> Path {
-        var path = Path()
-
-        path.move(to: CGPoint(x: rect.midX, y: rect.minY))
-        path.addLine(to: CGPoint(x: rect.minX, y: rect.maxY))
-        path.addLine(to: CGPoint(x: rect.maxX, y: rect.maxY))
-        path.addLine(to: CGPoint(x: rect.midX, y: rect.minY))
-        path.closeSubpath()
-
-        return path
-    }
-}
